Fix to return value of recursive delete
git-svn-id: svn://php-wax.com/svn/wxframework/trunk@916 0e7ba554-dfdc-446b-ac9a-b820edd20ad7

+++ wax/File.php
@@ -93,6 +93,7 @@ abstract class File {
 	}
 	
 	static function recursively_delete($item) {
+	  if(!is_file($item) && !is_dir($item)) return true;
 		if(is_file($item)) { unlink($item); return true; }
 		$iter = new RecursiveIteratorIterator(new RecursiveDirectoryIterator($item), 2);
 		foreach ( $iter as $file ) {
